Cancel order(s)

Cancel order(s)

Usage

cancel_order(orderIds = NULL, delay = 0, retries = 3)

Arguments

orderIds

A character vector of one or more which contains the order id(s) designated by KuCoin (required - default NULL).

delay

A numeric value to delay data request in milliseconds (optional - default 0).

retries

A numeric value to specify the number of retries in case of failure (optional - default 3).

Value

If success returns character vector of n order id(s); designated by KuCoin.

Details

This endpoint requires the Trade permission.

This API is restricted for each account, the request rate limit is 60 times/3s.

This interface is only for cancellation requests. The cancellation result needs to be obtained by querying the order status, you can use get_orders_by_id().

It is recommended that you DO NOT cancel the order until receiving the Open message, otherwise the order cannot be cancelled successfully.

—————

For more information see documentation: KuCoin - cancel-order

Examples

if (FALSE) { # \dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# set a limit order
order_id1 <- kucoin::post_kucoin_limit_order(
   symbol = "ETH/USDT",
   side = "sell",
   base_size = 1,
   price = 100000
)

order_id2 <- kucoin::post_kucoin_limit_order(
   symbol = "ETH/USDT",
   side = "sell",
   base_size = 1,
   price = 100000
)

# get order details
cancelled <- kucoin::cancel_order(c(order_id1, order_id2))

} # }