Get currencies' details

Get a currencies' details. This includes what chains are available for depositing; this function is useful for then generating a deposit address by use of [kucoin::get_deposit_address()].

| currency | name | full_name | precision | is_margin_enabled | is_debit_enabled | chain_name | chain | withdrawal_min_size | withdrawal_min_fee | is_withdraw_enabled | is_deposit_enabled | confirms | pre_confirms | contract_address | |———-|——|———–|———–|——————-|——————|————|——–|———————|——————–|———————|——————–|———-|————–|——————————————–| | BTC | BTC | Bitcoin | 8 | TRUE | TRUE | BTC | btc | 0.0008 | 0.0005 | TRUE | TRUE | 3 | 1 | | | BTC | BTC | Bitcoin | 8 | TRUE | TRUE | KCC | kcc | 0.0008 | 0.00002 | TRUE | TRUE | 20 | 20 | 0xfa93c12cd345c658bc4644d1d4e1b9615952258c | | BTC | BTC | Bitcoin | 8 | TRUE | TRUE | BTC-Segwit | bech32 | 0.0008 | 0.0005 | FALSE | TRUE | 2 | 2 | |

Usage

get_currency_details(currencies = NULL)

Arguments

currencies

A `character` vector to specify the currencies to get details for (required - default `NULL`).

Value

A `data.table` with currency information

Details

For more information see documentation: [KuCoin - get-currency-detail](https://docs.kucoin.com/#get-currency-detail-recommend)

Using v2 of the api.

See also

Examples

# get a currencies' details
kucoin::get_currency_details(c("BTC", "XMR"))
#>    currency   name full_name precision is_margin_enabled is_debit_enabled
#>      <char> <char>    <char>     <int>            <lgcl>           <lgcl>
#> 1:      BTC    BTC   Bitcoin         8              TRUE             TRUE
#> 2:      BTC    BTC   Bitcoin         8              TRUE             TRUE
#> 3:      BTC    BTC   Bitcoin         8              TRUE             TRUE
#> 4:      BTC    BTC   Bitcoin         8              TRUE             TRUE
#> 5:      XMR    XMR    Monero         8             FALSE             TRUE
#>           chain_name  chain withdrawal_min_size deposit_min_size
#>               <char> <char>              <char>           <char>
#> 1:               BTC    btc              0.0012           0.0002
#> 2: Lightning Network  btcln             0.00001          0.00001
#> 3:               KCC    kcc              0.0008             <NA>
#> 4:        BTC-Segwit bech32              0.0008           0.0002
#> 5:               XMR    xmr               0.007             <NA>
#>    withdraw_fee_rate withdraw_max_fee withdrawal_min_fee is_withdraw_enabled
#>               <char>           <lgcl>             <char>              <lgcl>
#> 1:                 0               NA             0.0006                TRUE
#> 2:                 0               NA           0.000015                TRUE
#> 3:                 0               NA            0.00002                TRUE
#> 4:                 0               NA             0.0005               FALSE
#> 5:                 0               NA              0.005                TRUE
#>    is_deposit_enabled confirms pre_confirms
#>                <lgcl>    <int>        <int>
#> 1:               TRUE        3            1
#> 2:               TRUE        1            1
#> 3:               TRUE       20           20
#> 4:               TRUE        2            2
#> 5:               TRUE       12           12
#>                              contract_address
#>                                        <char>
#> 1:                                           
#> 2:                                           
#> 3: 0xfa93c12cd345c658bc4644d1d4e1b9615952258c
#> 4:                                           
#> 5: