Get historical market data for symbols

Get historical market data for symbols

Usage

get_market_data(
  symbols = NULL,
  from = NULL,
  to = NULL,
  frequency = NULL,
  delay = 0,
  retries = 3
)

Arguments

symbols

A character vector of one or more pair symbol (required - default NULL).

from

A character with valid date/datetime format, or date/datetime object as a start of datetime range (required - default NULL).

to

A character with valid date/datetime format, or date/datetime object as an end of datetime range (required - default NULL).

frequency

A character vector of one which specify the frequency option, see details for further information (required - default NULL).

delay

A numeric value to delay data request in milliseconds (optional - default 0).

retries

A numeric value to specify the number of retries in case of failure (optional - default 3).

Value

A data.table with price data

Details

There are several supported frequencies:

"1 minute"

"3 minutes"

"5 minutes"

"15 minutes"

"30 minutes"

"1 hour"

"2 hours"

"4 hours"

"6 hours"

"8 hours"

"12 hours"

"1 day"

"1 week"

—————

For more information see documentation: KuCoin - get-klines

Examples

# get one symbol
kucoin::get_market_data(
    symbols = "BTC/USDT",
    from = "2022-11-05 00:00:00",
    to = "2022-11-06 00:00:00",
    frequency = "1 hour"
)
#>       symbol            datetime    open    high     low   close    volume
#>       <char>              <POSc>   <num>   <num>   <num>   <num>     <num>
#>  1: BTC/USDT 2022-11-05 00:00:00 21144.0 21456.0 21077.7 21436.8 1286.8395
#>  2: BTC/USDT 2022-11-05 01:00:00 21436.6 21455.2 21315.1 21364.5  631.8531
#>  3: BTC/USDT 2022-11-05 02:00:00 21364.4 21413.1 21342.9 21384.3  435.5556
#>  4: BTC/USDT 2022-11-05 03:00:00 21384.4 21480.0 21373.7 21445.1  736.9410
#>  5: BTC/USDT 2022-11-05 04:00:00 21445.0 21467.0 21363.7 21365.1  392.5327
#>  6: BTC/USDT 2022-11-05 05:00:00 21365.6 21430.8 21365.0 21391.0  412.2512
#>  7: BTC/USDT 2022-11-05 06:00:00 21391.0 21445.3 21372.8 21402.4  320.0829
#>  8: BTC/USDT 2022-11-05 07:00:00 21401.6 21414.8 21381.0 21385.6  282.8574
#>  9: BTC/USDT 2022-11-05 08:00:00 21385.7 21389.6 21230.7 21291.7  803.1178
#> 10: BTC/USDT 2022-11-05 09:00:00 21290.6 21324.7 21235.4 21289.2  431.6739
#> 11: BTC/USDT 2022-11-05 10:00:00 21289.1 21321.7 21279.6 21318.1  303.0455
#> 12: BTC/USDT 2022-11-05 11:00:00 21318.7 21425.1 21311.5 21402.6  639.8064
#> 13: BTC/USDT 2022-11-05 12:00:00 21402.8 21419.6 21289.2 21310.3  513.6069
#> 14: BTC/USDT 2022-11-05 13:00:00 21310.3 21325.1 21245.0 21268.7  434.5345
#> 15: BTC/USDT 2022-11-05 14:00:00 21269.2 21332.7 21210.1 21282.7  596.1384
#> 16: BTC/USDT 2022-11-05 15:00:00 21281.3 21358.2 21268.0 21298.6  536.4977
#> 17: BTC/USDT 2022-11-05 16:00:00 21299.3 21345.2 21298.1 21324.5  377.3549
#> 18: BTC/USDT 2022-11-05 17:00:00 21324.0 21335.0 21267.4 21318.0  386.4274
#> 19: BTC/USDT 2022-11-05 18:00:00 21318.0 21348.0 21260.3 21291.9  491.5955
#> 20: BTC/USDT 2022-11-05 19:00:00 21292.0 21344.3 21284.1 21331.2  351.5463
#> 21: BTC/USDT 2022-11-05 20:00:00 21331.3 21380.4 21320.4 21350.7  297.1821
#> 22: BTC/USDT 2022-11-05 21:00:00 21350.6 21359.0 21322.2 21334.9  174.3055
#> 23: BTC/USDT 2022-11-05 22:00:00 21334.9 21346.2 21287.1 21296.9  270.1990
#> 24: BTC/USDT 2022-11-05 23:00:00 21296.9 21315.8 21234.8 21297.8  421.1061
#>       symbol            datetime    open    high     low   close    volume
#>     turnover
#>        <num>
#>  1: 27384270
#>  2: 13496892
#>  3:  9307688
#>  4: 15785774
#>  5:  8410830
#>  6:  8822156
#>  7:  6851213
#>  8:  6052171
#>  9: 17112494
#> 10:  9187757
#> 11:  6455634
#> 12: 13683262
#> 13: 10960759
#> 14:  9249235
#> 15: 12679601
#> 16: 11433332
#> 17:  8045647
#> 18:  8234619
#> 19: 10471987
#> 20:  7493677
#> 21:  6342940
#> 22:  3719457
#> 23:  5759956
#> 24:  8959366
#>     turnover

# get multiple symbols
kucoin::get_market_data(
    symbols = c("BTC/USDT", "XMR/BTC", "KCS/USDT"),
    from = "2022-11-05 00:00:00",
    to = "2022-11-06 00:00:00",
    frequency = "1 hour"
)
#>       symbol            datetime        open        high         low
#>       <char>              <POSc>       <num>       <num>       <num>
#>  1: BTC/USDT 2022-11-05 00:00:00 2.11440e+04 2.14560e+04 2.10777e+04
#>  2: BTC/USDT 2022-11-05 01:00:00 2.14366e+04 2.14552e+04 2.13151e+04
#>  3: BTC/USDT 2022-11-05 02:00:00 2.13644e+04 2.14131e+04 2.13429e+04
#>  4: BTC/USDT 2022-11-05 03:00:00 2.13844e+04 2.14800e+04 2.13737e+04
#>  5: BTC/USDT 2022-11-05 04:00:00 2.14450e+04 2.14670e+04 2.13637e+04
#>  6: BTC/USDT 2022-11-05 05:00:00 2.13656e+04 2.14308e+04 2.13650e+04
#>  7: BTC/USDT 2022-11-05 06:00:00 2.13910e+04 2.14453e+04 2.13728e+04
#>  8: BTC/USDT 2022-11-05 07:00:00 2.14016e+04 2.14148e+04 2.13810e+04
#>  9: BTC/USDT 2022-11-05 08:00:00 2.13857e+04 2.13896e+04 2.12307e+04
#> 10: BTC/USDT 2022-11-05 09:00:00 2.12906e+04 2.13247e+04 2.12354e+04
#> 11: BTC/USDT 2022-11-05 10:00:00 2.12891e+04 2.13217e+04 2.12796e+04
#> 12: BTC/USDT 2022-11-05 11:00:00 2.13187e+04 2.14251e+04 2.13115e+04
#> 13: BTC/USDT 2022-11-05 12:00:00 2.14028e+04 2.14196e+04 2.12892e+04
#> 14: BTC/USDT 2022-11-05 13:00:00 2.13103e+04 2.13251e+04 2.12450e+04
#> 15: BTC/USDT 2022-11-05 14:00:00 2.12692e+04 2.13327e+04 2.12101e+04
#> 16: BTC/USDT 2022-11-05 15:00:00 2.12813e+04 2.13582e+04 2.12680e+04
#> 17: BTC/USDT 2022-11-05 16:00:00 2.12993e+04 2.13452e+04 2.12981e+04
#> 18: BTC/USDT 2022-11-05 17:00:00 2.13240e+04 2.13350e+04 2.12674e+04
#> 19: BTC/USDT 2022-11-05 18:00:00 2.13180e+04 2.13480e+04 2.12603e+04
#> 20: BTC/USDT 2022-11-05 19:00:00 2.12920e+04 2.13443e+04 2.12841e+04
#> 21: BTC/USDT 2022-11-05 20:00:00 2.13313e+04 2.13804e+04 2.13204e+04
#> 22: BTC/USDT 2022-11-05 21:00:00 2.13506e+04 2.13590e+04 2.13222e+04
#> 23: BTC/USDT 2022-11-05 22:00:00 2.13349e+04 2.13462e+04 2.12871e+04
#> 24: BTC/USDT 2022-11-05 23:00:00 2.12969e+04 2.13158e+04 2.12348e+04
#> 25:  XMR/BTC 2022-11-05 00:00:00 7.50000e-03 7.52100e-03 7.43300e-03
#> 26:  XMR/BTC 2022-11-05 01:00:00 7.49000e-03 7.52200e-03 7.46000e-03
#> 27:  XMR/BTC 2022-11-05 02:00:00 7.48100e-03 7.51800e-03 7.43500e-03
#> 28:  XMR/BTC 2022-11-05 03:00:00 7.50700e-03 7.51400e-03 7.44800e-03
#> 29:  XMR/BTC 2022-11-05 04:00:00 7.46800e-03 7.50700e-03 7.46500e-03
#> 30:  XMR/BTC 2022-11-05 05:00:00 7.50400e-03 7.52700e-03 7.46700e-03
#> 31:  XMR/BTC 2022-11-05 06:00:00 7.52100e-03 7.57000e-03 7.52100e-03
#> 32:  XMR/BTC 2022-11-05 07:00:00 7.56400e-03 7.56400e-03 7.47700e-03
#> 33:  XMR/BTC 2022-11-05 08:00:00 7.52700e-03 7.53200e-03 7.49100e-03
#> 34:  XMR/BTC 2022-11-05 09:00:00 7.51700e-03 7.53400e-03 7.49600e-03
#> 35:  XMR/BTC 2022-11-05 10:00:00 7.51700e-03 7.52500e-03 7.50300e-03
#> 36:  XMR/BTC 2022-11-05 11:00:00 7.50800e-03 7.54000e-03 7.49800e-03
#> 37:  XMR/BTC 2022-11-05 12:00:00 7.50700e-03 7.52900e-03 7.48500e-03
#> 38:  XMR/BTC 2022-11-05 13:00:00 7.51000e-03 7.52200e-03 7.47500e-03
#> 39:  XMR/BTC 2022-11-05 14:00:00 7.49600e-03 7.49600e-03 7.42800e-03
#> 40:  XMR/BTC 2022-11-05 15:00:00 7.44000e-03 7.46500e-03 7.42300e-03
#> 41:  XMR/BTC 2022-11-05 16:00:00 7.45800e-03 7.48400e-03 7.44700e-03
#> 42:  XMR/BTC 2022-11-05 17:00:00 7.47300e-03 7.48400e-03 7.44500e-03
#> 43:  XMR/BTC 2022-11-05 18:00:00 7.44900e-03 7.45100e-03 7.42000e-03
#> 44:  XMR/BTC 2022-11-05 19:00:00 7.43000e-03 7.44400e-03 7.40800e-03
#> 45:  XMR/BTC 2022-11-05 20:00:00 7.42100e-03 7.44100e-03 7.40500e-03
#> 46:  XMR/BTC 2022-11-05 21:00:00 7.41800e-03 7.44400e-03 7.41100e-03
#> 47:  XMR/BTC 2022-11-05 22:00:00 7.41500e-03 7.42800e-03 7.40100e-03
#> 48:  XMR/BTC 2022-11-05 23:00:00 7.42300e-03 7.44800e-03 7.37100e-03
#> 49: KCS/USDT 2022-11-05 00:00:00 1.04130e+01 1.04850e+01 1.01020e+01
#> 50: KCS/USDT 2022-11-05 01:00:00 1.01540e+01 1.02370e+01 1.01450e+01
#> 51: KCS/USDT 2022-11-05 02:00:00 1.02370e+01 1.02510e+01 1.02240e+01
#> 52: KCS/USDT 2022-11-05 03:00:00 1.02490e+01 1.02670e+01 1.02010e+01
#> 53: KCS/USDT 2022-11-05 04:00:00 1.02140e+01 1.02370e+01 1.02060e+01
#> 54: KCS/USDT 2022-11-05 05:00:00 1.02120e+01 1.02730e+01 1.02110e+01
#> 55: KCS/USDT 2022-11-05 06:00:00 1.02660e+01 1.02860e+01 1.02590e+01
#> 56: KCS/USDT 2022-11-05 07:00:00 1.02850e+01 1.02870e+01 1.02650e+01
#> 57: KCS/USDT 2022-11-05 08:00:00 1.02840e+01 1.02840e+01 1.01830e+01
#> 58: KCS/USDT 2022-11-05 09:00:00 1.01980e+01 1.02220e+01 1.01710e+01
#> 59: KCS/USDT 2022-11-05 10:00:00 1.01950e+01 1.02340e+01 1.01900e+01
#> 60: KCS/USDT 2022-11-05 11:00:00 1.02240e+01 1.02620e+01 1.02230e+01
#> 61: KCS/USDT 2022-11-05 12:00:00 1.02560e+01 1.02600e+01 1.02090e+01
#> 62: KCS/USDT 2022-11-05 13:00:00 1.02310e+01 1.02380e+01 1.01850e+01
#> 63: KCS/USDT 2022-11-05 14:00:00 1.02210e+01 1.02430e+01 1.01740e+01
#> 64: KCS/USDT 2022-11-05 15:00:00 1.01830e+01 1.02330e+01 1.01820e+01
#> 65: KCS/USDT 2022-11-05 16:00:00 1.02180e+01 1.02570e+01 1.02180e+01
#> 66: KCS/USDT 2022-11-05 17:00:00 1.02570e+01 1.02570e+01 1.02100e+01
#> 67: KCS/USDT 2022-11-05 18:00:00 1.02180e+01 1.02500e+01 1.01830e+01
#> 68: KCS/USDT 2022-11-05 19:00:00 1.01960e+01 1.02090e+01 1.01900e+01
#> 69: KCS/USDT 2022-11-05 20:00:00 1.02060e+01 1.02270e+01 1.02050e+01
#> 70: KCS/USDT 2022-11-05 21:00:00 1.02210e+01 1.02210e+01 1.01900e+01
#> 71: KCS/USDT 2022-11-05 22:00:00 1.02010e+01 1.02560e+01 1.02010e+01
#> 72: KCS/USDT 2022-11-05 23:00:00 1.02120e+01 1.02310e+01 1.01730e+01
#>       symbol            datetime        open        high         low
#>           close     volume     turnover
#>           <num>      <num>        <num>
#>  1: 2.14368e+04  1286.8395 2.738427e+07
#>  2: 2.13645e+04   631.8531 1.349689e+07
#>  3: 2.13843e+04   435.5556 9.307688e+06
#>  4: 2.14451e+04   736.9410 1.578577e+07
#>  5: 2.13651e+04   392.5327 8.410830e+06
#>  6: 2.13910e+04   412.2512 8.822156e+06
#>  7: 2.14024e+04   320.0829 6.851213e+06
#>  8: 2.13856e+04   282.8574 6.052171e+06
#>  9: 2.12917e+04   803.1178 1.711249e+07
#> 10: 2.12892e+04   431.6739 9.187757e+06
#> 11: 2.13181e+04   303.0455 6.455634e+06
#> 12: 2.14026e+04   639.8064 1.368326e+07
#> 13: 2.13103e+04   513.6069 1.096076e+07
#> 14: 2.12687e+04   434.5345 9.249235e+06
#> 15: 2.12827e+04   596.1384 1.267960e+07
#> 16: 2.12986e+04   536.4977 1.143333e+07
#> 17: 2.13245e+04   377.3549 8.045647e+06
#> 18: 2.13180e+04   386.4274 8.234619e+06
#> 19: 2.12919e+04   491.5955 1.047199e+07
#> 20: 2.13312e+04   351.5463 7.493677e+06
#> 21: 2.13507e+04   297.1821 6.342940e+06
#> 22: 2.13349e+04   174.3055 3.719457e+06
#> 23: 2.12969e+04   270.1990 5.759956e+06
#> 24: 2.12978e+04   421.1061 8.959366e+06
#> 25: 7.49300e-03   505.0832 3.779785e+00
#> 26: 7.48500e-03   193.2858 1.448284e+00
#> 27: 7.50000e-03   386.1058 2.889902e+00
#> 28: 7.46200e-03    76.6583 5.724171e-01
#> 29: 7.50300e-03    84.8136 6.355597e-01
#> 30: 7.52500e-03    39.8836 2.990697e-01
#> 31: 7.55000e-03   158.8947 1.200302e+00
#> 32: 7.52800e-03   268.1510 2.015025e+00
#> 33: 7.51700e-03    96.3912 7.248386e-01
#> 34: 7.51800e-03   243.6971 1.832038e+00
#> 35: 7.51300e-03    55.2685 4.152951e-01
#> 36: 7.50800e-03   160.3695 1.206822e+00
#> 37: 7.50900e-03   235.4337 1.768536e+00
#> 38: 7.49500e-03   129.5930 9.723318e-01
#> 39: 7.42800e-03   243.0590 1.813562e+00
#> 40: 7.45800e-03   645.9796 4.813401e+00
#> 41: 7.47500e-03   489.3374 3.652363e+00
#> 42: 7.45000e-03   759.8507 5.666950e+00
#> 43: 7.43400e-03   426.0155 3.167505e+00
#> 44: 7.42100e-03   481.2465 3.570506e+00
#> 45: 7.41900e-03   356.0630 2.639918e+00
#> 46: 7.41500e-03    92.4389 6.862592e-01
#> 47: 7.42000e-03   117.7360 8.725282e-01
#> 48: 7.39200e-03   546.8093 4.044374e+00
#> 49: 1.01540e+01 72511.8996 7.446657e+05
#> 50: 1.02360e+01 13658.4223 1.392494e+05
#> 51: 1.02490e+01  7411.1736 7.586906e+04
#> 52: 1.02130e+01  5186.1985 5.308258e+04
#> 53: 1.02110e+01  1789.5124 1.829325e+04
#> 54: 1.02650e+01  2094.1163 2.147240e+04
#> 55: 1.02850e+01  2140.9302 2.199944e+04
#> 56: 1.02840e+01  2233.3440 2.295634e+04
#> 57: 1.01980e+01  5959.9024 6.087437e+04
#> 58: 1.01940e+01  3294.9482 3.358959e+04
#> 59: 1.02240e+01  2445.1422 2.498673e+04
#> 60: 1.02560e+01  3943.3385 4.038433e+04
#> 61: 1.02320e+01  2515.1654 2.572816e+04
#> 62: 1.02210e+01  4800.5446 4.904731e+04
#> 63: 1.01830e+01  3570.0170 3.639994e+04
#> 64: 1.02190e+01  6617.3513 6.752109e+04
#> 65: 1.02570e+01  3510.4407 3.594175e+04
#> 66: 1.02170e+01  3009.9724 3.077825e+04
#> 67: 1.01950e+01  5227.9688 5.337147e+04
#> 68: 1.02080e+01  1866.1018 1.903293e+04
#> 69: 1.02210e+01  2503.4842 2.558072e+04
#> 70: 1.02000e+01  1628.0691 1.661188e+04
#> 71: 1.02120e+01  2743.6777 2.806546e+04
#> 72: 1.02210e+01  7536.3611 7.676274e+04
#>           close     volume     turnover