Get all order(s)

TODO: experimental.

Usage

get_orders_all(
  symbol = NULL,
  status = NULL,
  side = NULL,
  type = NULL,
  tradeType = "TRADE",
  from = NULL,
  to = NULL,
  retries = 3,
  delay = 0.1
)

Arguments

symbol

A character vector of one or more which contain symbol(s) of format "BTC/USDT" (optional - default NULL).

status

A character vector of one either "done" or "active" (optional - default NULL).

side

A character vector of one either "buy" or "sell" (optional - default NULL).

type

A character vector of one either "limit", "market", "limit_stop", "market_stop" (optional - default NULL).

tradeType

A character vector of one either "TRADE" (spot trading), "MARGIN_TRADE" (cross margin trading), "MARGIN_ISOLATED_TRADE" (isolated margin trading) (required - default TRADE).

from

A character with valid date/datetime format, or date/datetime object as a start of datetime range (optional - default NULL).

to

A character with valid date/datetime format, or date/datetime object as an end of datetime range (optional - default NULL).

retries

A numeric value to specify the number of retries in case of failure (optional - default 3).

delay

A numeric value to delay data request in milliseconds (optional - default 0.1).

Value

A data.table containing order details

Details

For more information see documentation: KuCoin - list-orders

Examples

if (FALSE) { # \dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

kucoin::get_orders_all()

} # }