Post a limit order

Post a limit order

Usage

submit_limit_order(
  symbol = NULL,
  side = NULL,
  base_size = NULL,
  price = NULL,
  timeInForce = "GTC",
  cancelAfter = NULL,
  postOnly = NULL,
  hidden = NULL,
  iceberg = NULL,
  visibleSize = NULL
)

Arguments

symbol

A `character` vector of one or more pair symbol (required - default `NULL`).

side

A `character` vector of one which specify the order side: `"buy"` or `"sell"` (required - default `NULL`).

base_size

A `numeric` vector of one determining the base size of the order; n units of the first currency in the pair (required - default `NULL`).

price

A `numeric` vector of one which specify the price of the order (required - default `NULL`).

timeInForce

A `character` vector of one specifying the time in force policy: `"GTC"` (Good Till Canceled), `"GTT"` (Good Till Time), `"IOC"` (Immediate Or Cancel), or `"FOK"` (Fill Or Kill) (optional - default `"GTC"`).

cancelAfter

A `numeric` vector of one specifying the number of seconds to wait before cancelling the order (optional - default `NULL`).

postOnly

A `logical` vector of one specifying whether the order is post only; invalid when `timeInForce` is `"IOC"` or `"FOK"` (optional - default `NULL`).

hidden

A `logical` vector of one specifying whether the order is hidden (optional - default `NULL`).

iceberg

A `logical` vector of one specifying whether the order is iceberg (optional - default `NULL`).

visibleSize

A `numeric` vector of one specifying the visible size of the iceberg order (optional - default `0`).

Value

If success returns `character` vector of one; order id designated by KuCoin.

Details

This API is restricted for each account, the request rate limit is 45 times/3s.

Currencies are traded in pairs. The first currency is called the base currency and the second currency is called the quote currency. So for example, BTC/USDT, means that the base currency is the BTC and the quote currency is the USDT.

This function returns the order ID set by KuCoin typically looks like this: "63810a330091a60001ceeb04". This can be used to get the status of the order. See the function [kucoin::get_orders_by_id()].

# ————— Time in force policies provide guarantees about the lifetime of an order. There are four policies: Good Till Canceled GTC, Good Till Time GTT, Immediate Or Cancel IOC, and Fill Or Kill FOK.

1. GTC Good Till Canceled orders remain open on the book until canceled. This is the default behavior if no policy is specified. 1. GTT Good Till Time orders remain open on the book until canceled or the allotted cancelAfter is depleted on the matching engine. GTT orders are guaranteed to cancel before any other order is processed after the cancelAfter seconds placed in order book. 1. IOC Immediate Or Cancel orders instantly cancel the remaining size of the limit order instead of opening it on the book. 1. FOK Fill Or Kill orders are rejected if the entire size cannot be matched.

Note that self trades belong to match as well. For market orders, using the “TimeInForce” parameter has no effect.

The post-only flag ensures that the trader always pays the maker fee and provides liquidity to the order book. If any part of the order is going to pay taker fee, the order will be fully rejected.

If a post only order will get executed immediately against the existing orders (except iceberg and hidden orders) in the market, the order will be cancelled.

For post only orders, it will get executed immediately against the iceberg orders and hidden orders in the market. Users placing the post only order will be charged the maker fees and the iceberg and hidden orders will be charged the taker fees.

# ————— For more information see the [KuCoin API documentation - new order](https://docs.kucoin.com/#place-a-new-order).

Examples

if (FALSE) { # \dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# check our balances
kucoin::get_account_balances()

## -------
# to avoid errors we will calculate our price and order size
# kucoin accepts these values in specific increments per ticker symbol

# get the market metadata
ticker <- "BTC/USDT"
price_per_coin <- 10000
amount_to_spend <- 300

metadata <- kucoin::get_market_metadata.deprecated()[symbol == ticker, ]
price_increment <- metadata[symbol == ticker, ]$price_increment
size_increment <- metadata[symbol == ticker, ]$base_increment

# calculate size and price
size <- floor(amount_to_spend / price_per_coin / size_increment) * size_increment
price <- floor(price_per_coin / price_increment) * price_increment

## ----
# submit order
message(stringr::str_interp('${ticker}: selling ${size} ${gsub("\\\\/.*", "", ticker)} @ ${price} ${gsub(".*\\\\/", "", ticker)}.'))
order_id <- kucoin::submit_limit_order(
    symbol = ticker,
    side = "sell",
    base_size = size,
    price = price
); order_id

## ----
# buying asset so recalculate
price_per_coin <- 0.0005
amount_to_spend <- 300

# calculate size and price
size <- floor(amount_to_spend / price_per_coin / size_increment) * size_increment
price <- floor(price_per_coin / price_increment) * price_increment

message(stringr::str_interp('${ticker}: buying ${size} ${gsub("\\\\/.*", "", ticker)} @ ${price} ${gsub(".*\\\\/", "", ticker)}.'))
order_id2 <- kucoin::submit_limit_order(
    symbol = ticker,
    side = "buy",
    base_size = size,
    price = price
); order_id2

kucoin::get_orders_by_id(c(order_id, order_id2))

} # }