Post a market order

Post a market order

Usage

submit_market_order(
  symbol = NULL,
  side = NULL,
  base_size = NULL,
  quote_size = NULL
)

Arguments

symbol

A `character` vector of one or more pair symbol (required - default `NULL`).

side

A `character` vector of one which specify the order side: `"buy"` or `"sell"` (required - default `NULL`).

base_size

A `numeric` vector of one determining the base size of the order; n units of the first currency in the pair (required or `quote_size` - default `NULL`).

quote_size

A `numeric` vector which specify the base or quote currency size; n units of the second currency in the pair (required or `base_size` - default `NULL`).

Value

If success returns `character` vector of one; order id designated by KuCoin.

Details

For more information see the [KuCoin API documentation - new order](https://docs.kucoin.com/#place-a-new-order).

This API is restricted for each account, the request rate limit is 45 times/3s.

Currencies are traded in pairs. The first currency is called the base currency and the second currency is called the quote currency. So for example, BTC/USDT, means that the base currency is the BTC and the quote currency is the USDT.

This function returns the order ID set by KuCoin typically looks like this: "63810a330091a60001ceeb04". This can be used to get the status of the order. See the function [kucoin::get_orders_by_id()].

Examples